# sample/service-user-center/docs/api/single.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user": {
      "post": {
        "operationId": "UserHttpController_insert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInsertReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInsertResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      },
      "get": {
        "operationId": "UserHttpController_selectByPage",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "telephone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emailVerify",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "telephoneVerify",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "accountType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userIcon",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registeTime",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pwdLastModifyTime",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSelectByPageResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "UserHttpController_selectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSelectByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      },
      "patch": {
        "operationId": "UserHttpController_updateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      },
      "delete": {
        "operationId": "UserHttpController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    },
    "/api/user/m": {
      "post": {
        "operationId": "UserHttpController_insertMulti",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInsertMultiResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    },
    "/api/user/m/{ids}": {
      "get": {
        "operationId": "UserHttpController_selectByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSelectByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      },
      "patch": {
        "operationId": "UserHttpController_updateByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      },
      "delete": {
        "operationId": "UserHttpController_deleteByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    },
    "/api/user/login": {
      "post": {
        "operationId": "UserHttpController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    },
    "/api/user/detail": {
      "post": {
        "operationId": "UserHttpController_detail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    },
    "/api/user/{id}/password": {
      "put": {
        "operationId": "UserHttpController_resetDefaultPwd",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BodyBooleanResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "User - 用户信息"
        ]
      }
    }
  },
  "info": {
    "title": "SAMPLE SERVICE USER CENTER",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PageResDto": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "number"
          },
          "pageNum": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "pageSize",
          "pageNum",
          "total"
        ]
      },
...
      "UpdateEffectInfo": {
        "type": "object",
        "properties": {
          "effectNum": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "effectNum"
        ]
      },
      "UpdateByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "DeleteEffectInfo": {
        "type": "object",
        "properties": {
          "effectNum": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "effectNum"
        ]
      },
      "DeleteByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
...
      "UpdateByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "DeleteByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
...
      "UserInsertReqDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "emailVerify": {
            "type": "number"
          },
          "telephoneVerify": {
            "type": "number"
          },
          "accountType": {
            "type": "number"
          },
          "userIcon": {
            "type": "string"
          },
          "registeTime": {
            "type": "number"
          },
          "pwdLastModifyTime": {
            "type": "number"
          }
        }
      },
      "UserInsertResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "emailVerify": {
            "type": "number"
          },
          "telephoneVerify": {
            "type": "number"
          },
          "accountType": {
            "type": "number"
          },
          "userIcon": {
            "type": "string"
          },
          "registeTime": {
            "type": "number"
          },
          "pwdLastModifyTime": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "username",
          "emailVerify",
          "telephoneVerify",
          "accountType",
          "registeTime",
          "pwdLastModifyTime"
        ]
      },
      "UserInsertResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UserInsertResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "UserSelectResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "emailVerify": {
            "type": "number"
          },
          "telephoneVerify": {
            "type": "number"
          },
          "accountType": {
            "type": "number"
          },
          "userIcon": {
            "type": "string"
          },
          "registeTime": {
            "type": "number"
          },
          "pwdLastModifyTime": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "username",
          "emailVerify",
          "telephoneVerify",
          "accountType",
          "registeTime",
          "pwdLastModifyTime"
        ]
      },
      "UserSelectByPageResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSelectResDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "page",
          "message"
        ]
      },
      "UserSelectByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSelectResDto"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "UserUpdateReqDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "emailVerify": {
            "type": "number"
          },
          "telephoneVerify": {
            "type": "number"
          },
          "accountType": {
            "type": "number"
          },
          "userIcon": {
            "type": "string"
          },
          "registeTime": {
            "type": "number"
          },
          "pwdLastModifyTime": {
            "type": "number"
          }
        }
      },
      "UserInsertMultiResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserInsertResDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "UserSelectByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSelectResDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "UserLoginReqDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UserLoginResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "UserDetailResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UserSelectResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "BodyBooleanResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      }
    }
  }
}